Added docs for how to import AType and EType (#168)
Every time I try to use Opaque normal types in a project I always have to clone the project and search for the import to remember how it's done. Let's add this to the docs to save me and probably others a lot of time trying to figure out how to import this.